Forced double in trig_helpers.

+++ src/trig_helper.h
@@ -8,8 +8,8 @@
 #define PI 3.1415926535897932384626433832795
 #define TWO_PI 6.283185307179586476925286766559
 
-#define rads(DEGS)  ( (DEGS) * (PI / 180.0) )
-#define degs(RADS)  ( (RADS) * (180.0 / PI) )
+#define rads(DEGS)  ( ((double)(DEGS)) * (PI / 180.0) )
+#define degs(RADS)  ( ((double)(RADS)) * (180.0 / PI) )
 
 #endif
 //end inclusion filter
